Move Lua test helpers out into separate (unpublished) package
